Fix the number of iterations generated by rRule and exRule
